requirements.txt
pandas>=1.3.0
openpyxl>=3.0.0
requests>=2.25.0
python-docx>=0.8.11
beautifulsoup4>=4.9.0
supabase>=1.0.0
python-dotenv>=0.19.0
